# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Space Case Cruises API
  version: 0.2.0
  description: |
    Admin-first community platform API for Space Case Cruises.
    Auth in local/dev uses one of these headers:
    - x-user-email
    - x-user-id
servers:
  - url: http://localhost:4000
security:
  - UserEmailAuth: []
  - UserIdAuth: []
components:
  securitySchemes:
    UserEmailAuth:
      type: apiKey
      in: header
      name: x-user-email
    UserIdAuth:
      type: apiKey
      in: header
      name: x-user-id
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
paths:
...
  /api/v1/admin/subgroups:
    post:
      tags: [SubGroups]
      summary: Create subgroup template (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, slug, code, default_cost_level]
              properties:
                name:
                  type: string
                slug:
                  type: string
                code:
                  type: string
                default_description:
                  type: string
                  nullable: true
                default_tile_image_url:
                  type: string
                  format: uri
                  nullable: true
                extension:
                  type: string
                  nullable: true
                default_cost_level:
                  type: integer
                  minimum: 0
                  maximum: 8
      responses:
        '201':
          description: Created subgroup
  /api/v1/admin/subgroups/{id}:
    patch:
      tags: [SubGroups]
      summary: Update subgroup template (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: string
                slug:
                  type: string
                code:
                  type: string
                default_description:
                  type: string
                  nullable: true
                default_tile_image_url:
                  type: string
                  format: uri
                  nullable: true
                extension:
                  type: string
                  nullable: true
                default_cost_level:
                  type: integer
                  minimum: 0
                  maximum: 8
      responses:
        '200':
          description: Updated subgroup
...